requirements.txt
streamlit
xarray
netcdf4
pandas
numpy
plotly
folium
streamlit-folium
psycopg2-binary
pgvector
sqlalchemy
groq
openai
langchain
langchain-community
chromadb
faiss-cpu
sentence-transformers
transformers
torch
scikit-learn
matplotlib
seaborn
requests
python-dotenv
aiohttp
pydantic
typing-extensions